requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
Pillow>=8.0.0
tqdm>=4.60.0
librosa>=0.8.0  # 可选，用于音频处理
tensorboard>=2.8.0  # TensorBoard可视化